[Vagrant] Update gitian-builder to latest version.
NOTE: It was not possible to keep both precise and lucid boxen without significant changes as one is currently supported by Canonical and the other is archived. Since the precise box is only used for Mac OS X builds, which was never completed anyway, we drop precise in favor of lucid. Future releases will be based on the upstream gitian build process.